# internal/openapi/spec.yaml
openapi: 3.1.0
info:
  title: System Wrangler API
  version: v1
  description: |
    HTTP API exposed by the System Wrangler backend. Every protected
    endpoint requires the `sw_session` cookie issued by
    `POST /api/auth/login` (or `POST /api/auth/setup` on a fresh
    install). State-changing requests must also carry the
    `X-Sw-Csrf: 1` header — the CSRF middleware rejects authenticated
    POST/PUT/PATCH/DELETE without it.

    This spec is the source of truth for the API surface. A drift test
    (`cmd/server/openapi_drift_test.go`) compares it to the routes
    registered on the production mux so the two cannot rot apart.
servers:
  - url: /
    description: Same-origin (the SPA and the API share the host).
...
security:
  - sessionCookie: []
paths:
...
  /api/auth/totp/verify:
    post:
      tags: [totp]
      summary: |
        Second-factor step. Consumes the challenge cookie issued by
        /api/auth/login when TOTP is enabled. Accepts either a TOTP
        code or a one-shot recovery code.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
                rememberDevice: { type: boolean }
      responses:
        "200":
          description: Login complete; session cookie issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Invalid challenge, wrong code, disabled account, or no challenge cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "423":
          description: Correct code on a locked account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountLocked"
        "429":
          description: Per-IP throttle exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: TOTP not configured (vault or stores missing).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/totp/setup:
    post:
      tags: [totp]
      summary: Begin TOTP enrollment — server returns a pending secret + QR.
      responses:
        "200":
          description: Pending secret, otpauth URI, and base64-encoded QR PNG.
          content:
            application/json:
              schema:
                type: object
                required: [secret, uri, qrPng]
                properties:
                  secret: { type: string, description: Base32-encoded TOTP secret. }
                  uri: { type: string, description: otpauth:// URI for QR encoding. }
                  qrPng: { type: string, description: Base64-encoded PNG of the QR code. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503":
          description: TOTP not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/totp/confirm:
    post:
      tags: [totp]
      summary: Confirm pending TOTP enrollment with a code; receive recovery codes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        "200":
          description: TOTP activated; one-time recovery codes returned.
          content:
            application/json:
              schema:
                type: object
                required: [recoveryCodes]
                properties:
                  recoveryCodes:
                    type: array
                    items: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: No session, or code did not verify.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Pending secret cannot be decrypted with the current master key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: TOTP not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/totp:
    delete:
      tags: [totp]
      summary: Disable TOTP for the caller (requires password + current code).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, code]
              properties:
                password: { type: string }
                code: { type: string }
      responses:
        "204": { description: TOTP disabled and trusted-device rows wiped. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422":
          description: TOTP secret cannot be decrypted; ask an admin to reset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: TOTP not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/devices:
    get:
      tags: [totp]
      summary: List trusted devices for the caller.
      responses:
        "200":
          description: Caller's trusted-device rows.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrustedDevice"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/devices/{id}:
    delete:
      tags: [totp]
      summary: Revoke one of the caller's trusted devices.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204": { description: Device revoked. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503":
          description: Trusted-device store not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sw_session
  responses:
    BadRequest:
      description: Request payload rejected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource does not exist (or is hidden from the caller's scope).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    AccountLocked:
      type: object
      required: [error, lockedUntil]
      properties:
        error: { type: string }
        lockedUntil:
          type: string
          format: date-time
          description: RFC3339 timestamp when the lock expires.
...
    User:
      type: object
      required: [id, username, email, theme, createdAt, totpEnabled, disabled, mustChangePassword]
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string }
        theme:
          type: string
          enum: ["", "light", "dark"]
        createdAt: { type: string, format: date-time }
        totpEnabled: { type: boolean }
        disabled: { type: boolean }
        disabledAt: { type: string, format: date-time }
        failedAttempts: { type: integer }
        lockedUntil: { type: string, format: date-time }
        mustChangePassword: { type: boolean }
    TrustedDevice:
      type: object
      required: [id, label, createdAt, lastUsedAt, expiresAt]
      properties:
        id: { type: string }
        label: { type: string }
        createdAt: { type: string, format: date-time }
        lastUsedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
...